Feat: 메인 페이지 상품 구분 및 할인 가격 통일
- 메인 페이지에서 추천상품과 최신상품 구분을 위해 상품 서비스 및 매퍼 수정
- 할인 가격 통일을 위해 카테고리 상세페이지의 스페셜 특가 부분 수정

+++ sgsgsg/src/main/java/com/sgsg/dra/domain/SpecialsProduct.java
@@ -18,7 +18,7 @@ public class SpecialsProduct {
     private String endDate;
     
  // 데이터베이스에 매핑되지 않는 임시 필드
-    private transient int salePrice;
+    private transient long  salePrice;
     
 	public Long getProductNum() {
 		return productNum;
@@ -80,10 +80,10 @@ public class SpecialsProduct {
 	public void setSaleCount(int saleCount) {
 		this.saleCount = saleCount;
 	}
-	public int getSalePrice() {
+	public long  getSalePrice() {
 		return salePrice;
 	}
-	public void setSalePrice(int salePrice) {
+	public void setSalePrice(long  salePrice) {
 		this.salePrice = salePrice;
 	}
 	public String getEndDate() {
